refactor: add shield rule files for posts and proposals
